refactor(MigrationTools.Host): restructure command line options and commands for better organization and clarity feat(MigrationTools.Host): introduce new command classes for migration execution and initialization fix(MigrationTools.Host): update command options to use Spectre.Console.Cli for better command line interface handling
refactor(MigrationTools.Host): remove ExecuteHostedService.cs and InitHostedService.cs, replace with MigrationService.cs for simplified service execution
feat(MigrationTools.Host): add Spectre.Console.Cli and YamlDotNet packages for enhanced CLI and YAML support
chore(MigrationTools.Host): remove unused CommandLineParser package reference from MigrationTools.Host.csproj
style(MigrationTools.Host): remove unused imports in MigrationToolHost.cs for cleaner code
