--- src/MigrationTools.Host/CommandLine/OptionsMode.cs
@@ -1,11 +1,0 @@
-﻿namespace MigrationTools.Host.CommandLine
-{
-    public enum OptionsMode
-    {
-        Reference = 0,
-        WorkItemTracking = 1,
-        Fullv2 = 2,
-        WorkItemTrackingv2 = 3,
-        Basic = 4
-    }
-}
